Refactored presetables to PresetManager
